+++ src/main/net.ts
@@ -12,9 +12,8 @@ import { getUseChineseAppServer } from '../common/persistedStore';
 import type { AppSpec } from '../ipc/apps';
 import { TokenInformation } from '../ipc/artifactoryToken';
 import { inRenderer as downloadProgress } from '../ipc/downloadProgress';
-import { inRenderer as proxyLogin } from '../ipc/proxyLogin';
 import { retrieveToken } from './artifactoryTokenStorage';
-import { storeProxyLoginRequest } from './proxyLogins';
+import { handleLoginRequest } from './proxyLogins';
 
 // Using the same session name as electron-updater, so that proxy credentials
 // (if required) only have to be sent once.
@@ -80,9 +79,66 @@ export const determineEffectiveUrl = (url: string) => {
     return effectiveUrl;
 };
 
+const isNordicArtifactoryUrl = (url: string) =>
+    /^https?:\/\/files\.nordicsemi\.(com|cn)\//.test(url);
+
+const shortNordicArtifactoryUrl = (url: string) => {
+    const longArtifactoryUrlRegex =
+        /^https:\/\/files\.nordicsemi\.(?<tld>com|cn)\/ui\/api\/v1\/download\?isNativeBrowsing=false&repoKey=(?<repo>[^&]+)&path=(?<path>.*)/;
+    const match = url.match(longArtifactoryUrlRegex);
+
+    if (match == null) return url;
+
+    const {
+        groups: { tld, repo, path },
+    } = match as { groups: Record<string, string> };
+
+    return `https://files.nordicsemi.${tld}/artifactory/${repo}/${path}`;
+};
+
+const getDownloadSize = (url: string, bearer?: string) =>
+    new Promise<number | undefined>(resolve => {
+        const effectiveUrl = isNordicArtifactoryUrl(url)
+            ? shortNordicArtifactoryUrl(url)
+            : url;
+
+        const request = net.request({
+            url: effectiveUrl,
+            method: 'HEAD',
+            session: session.fromPartition(NET_SESSION_NAME),
+        });
+        request.setHeader('pragma', 'no-cache');
+
+        if (bearer) {
+            request.setHeader('Authorization', `Bearer ${bearer}`);
+        }
+
+        request.on('response', response => {
+            const contentLength = response.headers['content-length'];
+
+            if (contentLength != null) {
+                resolve(Number(contentLength));
+                request.abort();
+            } else {
+                resolve(undefined);
+            }
+        });
+
+        request.on('login', handleLoginRequest);
+        request.on('error', () => resolve(undefined));
+        request.end();
+    });
+
 const downloadToBuffer = (url: string, options: DownloadOptions) =>
-    new Promise<Buffer>((resolve, reject) => {
+    // eslint-disable-next-line no-async-promise-executor
+    new Promise<Buffer>(async (resolve, reject) => {
         const effectiveUrl = determineEffectiveUrl(url);
+        const bearer = options.bearer ?? determineBearer(effectiveUrl);
+
+        const downloadSize =
+            options.app != null
+                ? await getDownloadSize(effectiveUrl, bearer)
+                : undefined;
 
         const request = net.request({
             url: effectiveUrl,
@@ -90,7 +146,6 @@ const downloadToBuffer = (url: string, options: DownloadOptions) =>
         });
         request.setHeader('pragma', 'no-cache');
 
-        const bearer = options.bearer ?? determineBearer(effectiveUrl);
         if (bearer) {
             request.setHeader('Authorization', `Bearer ${bearer}`);
         }
@@ -112,12 +167,11 @@ const downloadToBuffer = (url: string, options: DownloadOptions) =>
             const addToBuffer = (data: Buffer) => {
                 buffer.push(data);
             };
-            const downloadSize = Number(response.headers['content-length']);
             let progress = 0;
             response.on('data', data => {
                 addToBuffer(data);
                 progress += data.length;
-                if (options.app != null) {
+                if (options.app != null && downloadSize != null) {
                     reportInstallProgress(options.app, progress, downloadSize);
                 }
             });
@@ -131,18 +185,7 @@ const downloadToBuffer = (url: string, options: DownloadOptions) =>
             );
         });
         if (options.enableProxyLogin) {
-            request.on('login', (authInfo, callback) => {
-                if (
-                    authInfo.isProxy === false &&
-                    authInfo.host === 'files.nordicsemi.com'
-                ) {
-                    callback();
-                    return;
-                }
-
-                const requestId = storeProxyLoginRequest(callback);
-                proxyLogin.requestProxyLogin(requestId, authInfo);
-            });
+            request.on('login', handleLoginRequest);
         }
         request.on('error', error =>
             reject(