+++ src/main/bundledJlink.js
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-module.exports = 'V8.66';
+module.exports = 'V8.76';