+++ Changelog.md
@@ -2,9 +2,10 @@
 
 ### Changed
 
--   Load updates to nRF Connect for Desktop to a new version from our own CDN
-    (not from GitHub, as before): Usually use `files.nordicsemi.com`. Use
-    `files.nordicsemi.cn` if users choose to use the Chinese server.
+-   Logic for loading updates to nRF Connect for Desktop to a new version. The
+    mechanism is now based on CDN and not GitHub. In most cases, it uses
+    `files.nordicsemi.com`. If users choose to use the Chinese server, it uses
+    `files.nordicsemi.cn`.
 
 ## 5.2.0 - 2025-05-22